Fix text2image's --only_extract_font_properties mode to print variant kerning spacing
The ok_count == 2 part of the test always failed, which prevented
the kerning variations from ever being printed. As far as I can tell
it wasn't doing anything useful, but my C++ is (still) rusty, so
some light checking would be wise.

Also updates the function description to be in line with the
current output