lstmeval: Improve output by ensuring 'Truth:' text is encoded the same way as OCR output
This ensures that transformations like unicode normalisation are done on
the truth output as well as the OCR output, so that you can compare
the two properly.

Before this a perfect OCR result could show different lines for Truth and
OCR if the OCR output included characters that were normalised.